refactor(frontend): move WalletConnect reject function to lib
