+++ src/frontend/src/eth/services/wallet-connect.services.ts
@@ -14,27 +14,20 @@ import {
 } from '$lib/constants/analytics.contants';
 import { ProgressStepsSend, ProgressStepsSign } from '$lib/enums/progress-steps';
 import { trackEvent } from '$lib/services/analytics.services';
+import {
+	execute,
+	type WalletConnectCallBackParams,
+	type WalletConnectExecuteParams
+} from '$lib/services/wallet-connect.services';
 import { authStore } from '$lib/stores/auth.store';
-import { busy } from '$lib/stores/busy.store';
 import { i18n } from '$lib/stores/i18n.store';
-import { toastsError, toastsShow } from '$lib/stores/toasts.store';
+import { toastsError } from '$lib/stores/toasts.store';
 import type { OptionEthAddress } from '$lib/types/address';
 import type { ResultSuccess } from '$lib/types/utils';
-import type { OptionWalletConnectListener, WalletConnectListener } from '$lib/types/wallet-connect';
+import type { OptionWalletConnectListener } from '$lib/types/wallet-connect';
 import { isNullish, nonNullish } from '@dfinity/utils';
 import { BigNumber } from '@ethersproject/bignumber';
-import { getSdkError } from '@walletconnect/utils';
-import type { Web3WalletTypes } from '@walletconnect/web3wallet';
 import { get } from 'svelte/store';
-
-export interface WalletConnectCallBackParams {
-	request: Web3WalletTypes.SessionRequest;
-	listener: WalletConnectListener;
-}
-
-export type WalletConnectExecuteParams = Pick<WalletConnectCallBackParams, 'request'> & {
-	listener: OptionWalletConnectListener;
-};
 
 export type WalletConnectSendParams = WalletConnectExecuteParams & {
 	listener: OptionWalletConnectListener;
@@ -49,28 +42,6 @@ export type WalletConnectSignMessageParams = WalletConnectExecuteParams & {
 	modalNext: () => void;
 	progress: (step: ProgressStepsSign) => void;
 };
-
-export const reject = (params: WalletConnectExecuteParams): Promise<ResultSuccess> =>
-	execute({
-		params,
-		callback: async ({
-			request,
-			listener
-		}: WalletConnectCallBackParams): Promise<ResultSuccess> => {
-			busy.start();
-
-			const { id, topic } = request;
-
-			try {
-				await listener.rejectRequest({ topic, id, error: getSdkError('USER_REJECTED') });
-
-				return { success: true };
-			} finally {
-				busy.stop();
-			}
-		},
-		toastMsg: get(i18n).wallet_connect.error.request_rejected
-	});
 
 export const send = ({
 	address,
@@ -282,55 +253,3 @@ export const signMessage = ({
 		},
 		toastMsg: get(i18n).wallet_connect.info.sign_executed
 	});
-
-const execute = async ({
-	params: { request, listener },
-	callback,
-	toastMsg
-}: {
-	params: WalletConnectExecuteParams;
-	callback: (params: WalletConnectCallBackParams) => Promise<ResultSuccess>;
-	toastMsg: string;
-}): Promise<ResultSuccess> => {
-	const {
-		wallet_connect: {
-			error: { no_connection_opened, request_not_defined, unexpected_processing_request }
-		}
-	} = get(i18n);
-
-	if (isNullish(listener)) {
-		toastsError({
-			msg: { text: no_connection_opened }
-		});
-		return { success: false };
-	}
-
-	if (isNullish(request)) {
-		toastsError({
-			msg: { text: request_not_defined }
-		});
-		return { success: false };
-	}
-
-	try {
-		const { success, err } = await callback({ request, listener });
-
-		if (!success) {
-			return { success, err };
-		}
-
-		toastsShow({
-			text: toastMsg,
-			level: 'info',
-			duration: 2000
-		});
-	} catch (err: unknown) {
-		toastsError({
-			msg: { text: unexpected_processing_request },
-			err
-		});
-		return { success: false, err };
-	}
-
-	return { success: true };
-};
